Rename school rule files
