+++ apps/server/src/shared/domain/rules/legacy-school.rule.ts
